Added the PostPolicy methods

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Models\User;
+use Illuminate\Contracts\Session\Session;
 use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
@@ -16,9 +17,13 @@ class PostController extends Controller
      */
     public function index()
     {
-        //
-        $posts=Post::all();
+        // to show posts that were created by the logged in user.
+        // $id = Auth::id(); 
+        // $posts = Post::where("user_id", "=", $id)->get(); 
+        
+        // gets all posts
 
+        $posts=Post::all(); 
         return view('admin.posts.index',['posts'=>$posts]);
     }
 
@@ -40,26 +45,27 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
+
     {
         //validate input
-        $inputs=request()->validate([
+        $request->validate([
             'title'=>'required|min:2|max:255',
             'user_id'=>'required',
             'text_image'=>'file',
             'body'=>'required'
         ]);
-        if(request('text_image')){
-            $inputs['text_image']=$request->text_image;
+        if($request->hasFile('text_image')){
+            $img=$request->text_image;
             //rename image
-            $reImg=time().'.'.$inputs['text_image']->getClientOriginalExtension();
+            $reImg=time().'.'.$img->getClientOriginalExtension();
             //save image to path
             $dest=public_path('/images');
             //move file
-            $inputs['text_image']->move($dest,$reImg);
-
-
-
-        // }
+            $img->move($dest,$reImg);
+        }
+        else{
+            $reImg="nothing.png";
+        }
 
         $post=new Post();
 
@@ -70,9 +76,9 @@ class PostController extends Controller
 
         $post->save();
 
-        return redirect('/admin')->with('msg','Data updates successfully');
+        return redirect('/admin/post/index')->with('msg','Post was created successfully');
     }
-}
+
 
     /**
      * Display the specified resource.
@@ -96,6 +102,11 @@ class PostController extends Controller
     public function edit($id)
     {
         //
+        $post=Post::findorFail($id);
+        $this->authorize('view',$post); //this controller has utilised the postpolicy rules
+        return view('admin.posts.edit',['post'=>$post]);
+        
+
     }
 
     /**
@@ -108,6 +119,37 @@ class PostController extends Controller
     public function update(Request $request, $id)
     {
         //
+        $request->validate([
+            'title'=>'required|min:2|max:255',
+            'user_id'=>'required',
+            'text_image'=>'file',
+            'body'=>'required'
+        ]);
+        if($request->hasFile('text_image')){
+            $img=$request->text_image;
+            //rename image
+            $reImg=time().'.'.$img->getClientOriginalExtension();
+            //save image to path
+            $dest=public_path('/images');
+            //move file
+            $img->move($dest,$reImg);
+        }
+        else{
+            $reImg="nothing.png";
+        }
+        $post=new Post();
+
+        $post['title']=$request->title;
+        $post['user_id']=$request->user_id;
+        $post['text_image']=$reImg;
+        $post['body']=$request->body;
+
+        // $this->authorize('update',$post);
+        $post->update();
+
+        return redirect('/admin/post/index')->with('update_msg','post updated successfully');
+
+        
     }
 
     /**
@@ -118,10 +160,16 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
-        //
+        //hard coded not utilised the policy
+        
         $post=Post::findorFail($id);
-        $post->delete();
+        if(Auth::user()->id !== $post->user_id){
+            // what happens when an un-authenticated user tries to delete a post.
 
-        return redirect('/admin');
+        }
+        else{
+            $post->delete();
+            return redirect('/admin')->with('deletemsg','Post deleted successfully');
+        }       
     }
 }
